pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "greenlit"
version = "0.1.0"
description = "Structure prompts before you burn tokens"
requires-python = ">=3.11"
dependencies = ["rich>=13.0"]

[project.optional-dependencies]
templates = ["pyyaml>=6.0"]

[project.scripts]
greenlit = "greenlit.cli:main"

[project.urls]
Repository = "https://github.com/abali2509/greenlit"

[tool.hatch.build.targets.wheel]
packages = ["src/greenlit"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
